Revert "Added solution to Q6, combined Logging with Utilities"
